requirements.txt
streamlit
pinecone-client
xata
python-dotenv
google-generativeai
